chore: move props utils to pwt
