+++ packages/pluggableWidgets/bubble-chart-web/src/BubbleChart.editorConfig.ts
@@ -1,14 +1,13 @@
+import { ContainerProps, ImageProps, StructurePreviewProps } from "@mendix/piw-utils-internal";
 import {
-    ContainerProps,
     hideNestedPropertiesIn,
     hidePropertiesIn,
     hidePropertyIn,
-    ImageProps,
     Problem,
     Properties,
-    StructurePreviewProps,
     transformGroupsIntoTabs
-} from "@mendix/piw-utils-internal";
+} from "@mendix/pluggable-widgets-tools";
+
 import { BubbleChartPreviewProps } from "../typings/BubbleChartProps";
 
 import BubbleChartLightSvg from "./assets/BubbleChart.light.svg";
