+++ packages/pluggableWidgets/column-chart-native/src/ColumnChart.editorConfig.ts
@@ -1,4 +1,5 @@
-import { hideNestedPropertiesIn, Problem, Properties, StructurePreviewProps } from "@mendix/piw-utils-internal";
+import { StructurePreviewProps } from "@mendix/piw-utils-internal";
+import { hideNestedPropertiesIn, Problem, Properties } from "@mendix/pluggable-widgets-tools";
 
 import { ColumnChartPreviewProps } from "../typings/ColumnChartProps";
 import columnChartGroupedSvgDark from "./assets/ColumnChart.Grouped.dark.svg";
