+++ packages/pluggableWidgets/datagrid-number-filter-web/src/DatagridNumberFilter.editorConfig.ts
@@ -8,14 +8,11 @@ import {
     greaterThanEqualIconDark,
     greaterThanIcon,
     greaterThanIconDark,
-    hidePropertiesIn,
-    hidePropertyIn,
     ImageProps,
     notEmptyIcon,
     notEmptyIconDark,
     notEqualIcon,
     notEqualIconDark,
-    Properties,
     smallerThanEqualIcon,
     smallerThanEqualIconDark,
     smallerThanIcon,
@@ -23,6 +20,8 @@ import {
     StructurePreviewProps,
     TextProps
 } from "@mendix/piw-utils-internal";
+import { hidePropertiesIn, hidePropertyIn, Properties } from "@mendix/pluggable-widgets-tools";
+
 import { DatagridNumberFilterPreviewProps, DefaultFilterEnum } from "../typings/DatagridNumberFilterProps";
 
 export function getProperties(
