+++ packages/pluggableWidgets/heatmap-chart-web/src/HeatMap.editorConfig.ts
@@ -1,12 +1,6 @@
-import {
-    hidePropertiesIn,
-    Properties,
-    StructurePreviewProps,
-    transformGroupsIntoTabs,
-    ImageProps,
-    ContainerProps,
-    moveProperty
-} from "@mendix/piw-utils-internal";
+import { StructurePreviewProps, ImageProps, ContainerProps } from "@mendix/piw-utils-internal";
+import { hidePropertiesIn, moveProperty, Properties, transformGroupsIntoTabs } from "@mendix/pluggable-widgets-tools";
+
 import { HeatMapPreviewProps } from "../typings/HeatMapProps";
 
 import HeatMapDark from "./assets/HeatMap.dark.svg";
