+++ packages/pluggableWidgets/line-chart-native/src/LineChart.editorConfig.ts
@@ -1,4 +1,5 @@
-import { StructurePreviewProps, hideNestedPropertiesIn, Problem, Properties } from "@mendix/piw-utils-internal";
+import { StructurePreviewProps } from "@mendix/piw-utils-internal";
+import { hideNestedPropertiesIn, Problem, Properties } from "@mendix/pluggable-widgets-tools";
 
 import lineChartSvgDark from "./assets/LineChart.dark.svg";
 import lineChartSvgLight from "./assets/LineChart.light.svg";
