+++ packages/pluggableWidgets/timeline-web/src/Timeline.editorConfig.ts
@@ -1,10 +1,6 @@
-import {
-    hidePropertiesIn,
-    hidePropertyIn,
-    Problem,
-    Properties,
-    StructurePreviewProps
-} from "@mendix/piw-utils-internal";
+import { StructurePreviewProps } from "@mendix/piw-utils-internal";
+import { hidePropertiesIn, hidePropertyIn, Problem, Properties } from "@mendix/pluggable-widgets-tools";
+
 import { TimelinePreviewProps } from "../typings/TimelineProps";
 import lineAndDotSVG from "./assets/lineAndDot.svg";
 import lineAndDotSVGDark from "./assets/lineAndDot-dark.svg";
