+++ packages/tools/piw-utils-internal/src/utils/index.ts
@@ -1,4 +1,3 @@
-export * from "./PageEditorUtils";
 export * from "./StylingPropertyUtils";
 export { useScheduleUpdateOnce } from "./ReactLifecycleUtils";
 export * from "./AttributeValueUtils";
