+++ packages/tools/pluggable-widgets-tools/src/index.ts
@@ -1,3 +1,5 @@
 export * from "./common";
 export * from "./native/common";
 export * from "./web/common";
+export * from "./utils/typings";
+export * from "./utils";
