+++ packages/tools/pluggable-widgets-tools/src/utils/__tests__/PageEditorUtils.spec.ts
@@ -1,4 +1,4 @@
-import { Properties } from "../../typings/PageEditor";
+import { Properties } from "../typings";
 import { moveProperty } from "../PageEditorUtils";
 
 describe("The PageEditorUtils", () => {
