+++ packages/tools/pluggable-widgets-tools/src/utils/typings/PageEditor.ts
@@ -1,0 +1,30 @@
+export type Properties = PropertyGroup[];
+
+export type PropertyGroup = {
+    caption: string;
+    propertyGroups?: PropertyGroup[];
+    properties?: Property[];
+};
+
+export type Property = {
+    key: string;
+    caption: string;
+    description?: string;
+    objectHeaders?: string[]; // used for customizing object grids
+    objects?: ObjectProperties[];
+    properties?: Properties[]; // Property needs to remain here for compatibility with Studio Pro < 8.12
+};
+
+export type ObjectProperties = {
+    properties: PropertyGroup[];
+    captions?: string[]; // used for customizing object grids
+};
+
+export type Problem = {
+    property?: string; // key of the property, at which the problem exists
+    severity?: "error" | "warning" | "deprecation"; // default = "error"
+    message: string; // description of the problem
+    studioMessage?: string; // studio-specific message, defaults to message
+    url?: string; // link with more information about the problem
+    studioUrl?: string; // studio-specific link
+};
