+++ packages/tools/pluggable-widgets-tools/src/utils/typings/index.ts
@@ -1,0 +1,1 @@
+export * from "./PageEditor";
